requirements_dev.txt
coverage==4.0.3
isort==4.2.5
pep8==1.7.0
pytest==2.9.1
pytest-cov==2.2.1
pytest-django==2.9.1
pytest-isort==0.1.0
pytest_mock==0.11.0
pytest-pep8==1.0.6
pytest-pythonpath==0.7
pytest-sugar==0.7.1
tox==2.3.1
